refactor(pdf): consolidate on MVP engine; remove legacy stack
The MVP engine in src/utils/pdf_mvp/ is now the only PDF stack. The legacy
src/utils/pdf/ directory is deleted. Files still needed by the JPG/canvas
exporter (pdfLayout.js, fonts.js) are relocated into src/utils/media/ and
renamed to make their JPG-only role explicit:

  src/utils/pdf/pdfLayout.js -> src/utils/media/jpgPlanner.js
  src/utils/pdf/fonts.js     -> src/utils/media/canvasFonts.js
  src/utils/pdf/footer.js    -> src/utils/pdf_mvp/footer.js

Page-level callers (Bundle, Setlist, Songbook, SongView) now dynamic-import
src/utils/pdf_mvp directly. normalizeSongInput is dropped from the three
PDF-only pages (the MVP engine accepts the lyricsBlocks shape natively); it
is preserved on SongViewPage where the JPG flow still needs it.

Legacy planner tests (src/utils/pdf/__tests__/) and legacy-only tests in
src/__tests__/ (pdf.content, pdf.sectionsFromSong, planSongLayout.compare)
are deleted along with the planner they covered. The MVP test suite
(npm run test:mvp) and the JPG planner test continue to pass.
