--- src/utils/pdf/__tests__/fixtures/holy_forever.chordpro
@@ -1,91 +1,0 @@
-{title: Holy Forever}
-{key: C}
-{authors: Chris Tomlin}
-{country: USA}
-{tags: Slow, Contemporary}
-{youtube: IkHgxKemCRk}
-{mp3: }
-{pptx: }
-
-Verse 1
-A [C]thousand generations
-[F]Falling down in [C]worship
-To [Am]sing the song of [G]ages to the [F]Lamb
-And [C]all who’ve gone before us and
-[F]All who will [C]believe
-Will [Am]sing the song of [G]ages to the [F]Lamb
-
-Pre-Chorus
-Your [F]name is the highest
-Your [G]name is the greatest
-Your [Am]name stands above them [F]all
-All [F]thrones and dominions
-All [G]powers and positions
-Your [Am]name stands above them [Dm]all
-
-Chorus
-And the angels [F]cry, [G]Holy
-All creation [C]cries, [Am]Holy
-You are lifted [Dm]high, [G]Holy
-Holy [C]forever
-
-Hear your people [F]sing, [G]Holy
-To the King of [C]Kings, [Am]Holy
-You will always [Dm]be, [G]Holy
-Holy [C]forever
-
-Verse 2
-[C]If you’ve been forgiven and
-[F]If you’ve been re[C]deemed
-[Am]Sing the song [C]forever to the [F]Lamb
-[C]If you walk in freedom and [F]if you bear His [C]name
-[Am]Sing the song [G]forever to the [F]Lamb
-We’ll [Am]sing the song for[G]ever and [F]amen
-
-Tag
-You will always [Dm]be, Ho[G]ly
-Holy for[C]ever
-=======
-
-{verse: 1}
-[C]A thousand generations
-[F]Falling down in [C]worship
-[Am][G][F]To sing the song of ages to the Lamb
-[C]And all who’ve gone before us and
-[F]All who will [C]believe
-[Am][G][F]Will sing the song of ages to the Lamb
-{end_of_verse}
-
-{pre-chorus}
-[F]Your name is the highest
-[G]Your name is the greatest
-[Am][F]Your name stands above them all
-[F]All thrones and dominions
-[G]All powers and positions
-[Am][Dm]Your name stands above them all
-{end_of_pre-chorus}
-
-{chorus}
-[F][G]And the angels cry, Holy
-[C][Am]All creation cries, Holy
-[Dm][G]You are lifted high, Holy
-[C]Holy forever
-[F][G]Hear your people sing, Holy
-[C][Am]To the King of Kings, Holy
-[Dm][G]You will always be, Holy
-[C]Holy forever
-{end_of_chorus}
-
-{verse: 2}
-[C]If you’ve been forgiven and
-[F]If you’ve been re[C]deemed
-[Am][C][F]Sing the song forever to the Lamb
-[C][F][C]If you walk in freedom and if you bear His name
-[Am][G][F]Sing the song forever to the Lamb
-[Am][G][F]We’ll sing the song forever and amen
-{end_of_verse}
-
-{tag}
-[Dm][G]You will always be, Holy
-[C]Holy forever
-{end_of_tag}
