--- src/utils/pdf/__tests__/planner.hints.test.ts
@@ -1,43 +1,0 @@
-import { describe, it, expect } from 'vitest'
-import { chooseBestLayout, normalizeSongInput } from '../pdfLayout'
-
-const case3 = `
-{start_of_verse: Verse 1}
-[C]a
-[C]b
-{end_of_verse}
-{start_of_verse: Verse 2}
-[C]c
-[C]d
-{end_of_verse}
-{start_of_chorus}
-[C]e
-[C]f
-{end_of_chorus}
-{start_of_verse: Verse 3}
-[C]g
-[C]h
-{end_of_verse}
-`
-
-describe('Planner single-page priorities & hints', () => {
-  it('prefers 2-col single page over spilling to page 2 (Case 3)', async () => {
-    const mm = (pt: number) => (text: string) => (text ? text.length * (pt * 0.6) : 0)
-    const { plan } = chooseBestLayout(normalizeSongInput(case3) as any, {}, mm, mm)
-    expect(plan.layout.pages.length).toBe(1)
-    expect([1, 2]).toContain(plan.columns)
-    expect(plan.lyricSizePt).toBeGreaterThanOrEqual(12)
-  })
-
-  it('2-col tiny second column guard prefers 1-col at same size', async () => {
-    const tinyTail = `
-{start_of_verse} [C]aaa {end_of_verse}
-{start_of_verse} [C]bbb {end_of_verse}
-{start_of_verse} [C]ccc {end_of_verse}
-`
-    const mm = (pt: number) => (text: string) => (text ? text.length * (pt * 0.6) : 0)
-    const { plan } = chooseBestLayout(normalizeSongInput(tinyTail) as any, {}, mm, mm)
-    expect(plan.layout.pages.length).toBe(1)
-    expect(plan.columns).toBe(1)
-  })
-})
