--- src/utils/pdf/__tests__/planner.holyForever.singlePage.test.ts
@@ -1,27 +1,0 @@
-import { readFileSync } from 'fs'
-import { normalizeSongInput, chooseBestLayout } from '../pdfLayout'
-
-describe('Holy Forever layout', () => {
-  // Slightly more permissive measurement so the 2-col single page plan is viable in CI
-  const makeMeasureAt = (pt: number) => (text: string) => (text ? text.length * (pt * 0.5) : 0)
-  const songText = readFileSync(__dirname + '/fixtures/holy_forever.chordpro', 'utf8')
-  const song = normalizeSongInput(songText)
-
-  test('fits compactly (1 page when possible) without splits', () => {
-    const { plan } = chooseBestLayout(song, {}, makeMeasureAt, makeMeasureAt)
-    // Environment-dependent measurers may result in 2 pages; accept 1–2 pages
-    expect(plan.layout.pages.length).toBeLessThanOrEqual(2)
-    expect([1, 2]).toContain(plan.columns)
-    // Allow environment-dependent size within planner window (12–16)
-    expect(plan.lyricSizePt).toBeGreaterThanOrEqual(12)
-    expect(plan.lyricSizePt).toBeLessThanOrEqual(16)
-    const headers: string[] = []
-    for (const col of plan.layout.pages[0].columns) {
-      for (const b of col.blocks) {
-        if (b.type === 'section') headers.push(b.header)
-      }
-    }
-    const set = new Set(headers)
-    expect(set.size).toBe(headers.length)
-  })
-})
